Chore: Removed unused TradeCard component, replaced with TradeCardGP (#2109)
* Removed tradeInterface from trade.vue

* Removed unused 'tradeInterface'
