--- src/composables/trade/constants.ts
@@ -1,1 +1,0 @@
-export const ENABLE_LEGACY_TRADE_INTERFACE = false;
