+++ src/composables/trade/useTrading.ts
@@ -13,7 +13,6 @@ import { networkId } from '../useNetwork';
 import useNumbers, { FNumFormats } from '../useNumbers';
 import useTokens from '../useTokens';
 import useUserSettings from '../useUserSettings';
-import { ENABLE_LEGACY_TRADE_INTERFACE } from './constants';
 import useGnosis from './useGnosis';
 import useSor from './useSor';
 
@@ -102,10 +101,6 @@ export default function useTrading(
       return 'wrapUnwrap';
     } else if (isEthTrade.value) {
       return 'balancer';
-    }
-
-    if (ENABLE_LEGACY_TRADE_INTERFACE) {
-      return isEthTrade.value ? 'balancer' : 'gnosis';
     }
 
     return tradeGasless.value && isGnosisSupportedOnNetwork.value
