+++ src/store/modules/app.ts
@@ -2,16 +2,10 @@ import LS_KEYS from '@/constants/local-storage.keys';
 import { lsGet, lsSet } from '@/lib/utils';
 import i18n from '@/plugins/i18n';
 
-export enum TradeInterface {
-  GNOSIS = 'gnosis',
-  BALANCER = 'balancer',
-}
-
 export interface AppState {
   loading: boolean;
   modalOpen: boolean;
   locale: string;
-  tradeInterface: TradeInterface;
   transactionDeadline: number;
 }
 
@@ -20,7 +14,6 @@ const state: AppState = {
   modalOpen: false,
   locale: lsGet(LS_KEYS.App.Locale, 'en-US'),
   transactionDeadline: lsGet(LS_KEYS.App.TradeDeadline, 20), // minutes
-  tradeInterface: lsGet(LS_KEYS.App.TradeInterface, TradeInterface.BALANCER),
 };
 
 const actions = {
@@ -52,14 +45,6 @@ const mutations = {
     state.locale = locale;
     lsSet(LS_KEYS.App.Locale, locale);
     i18n.global.locale = locale;
-  },
-
-  setTradeInterface(
-    state: AppState,
-    tradeInterface: AppState['tradeInterface']
-  ) {
-    state.tradeInterface = tradeInterface;
-    lsSet(LS_KEYS.App.TradeInterface, state.tradeInterface);
   },
 
   setTransactionDeadline(
